updates on part 1 and part 2

+++ part_02/Exercise_07.java
@@ -18,8 +18,8 @@ public class Exercise_07 {
         Scanner userInput = new Scanner(System.in);
         System.out.print("Enter numbers in minutes : ");
 
-        int numbersInMinutes = userInput.nextInt();
-        int numberInDays = numbersInMinutes / 1440;
+        double numbersInMinutes = userInput.nextDouble();
+        double numberInDays = numbersInMinutes / 1440;
         double numberInYears = numberInDays / 365;
 
         System.out.println("In Days : " + numberInDays);
